+++ Profile/Shopware/Premapping/PaymentMethodReader.php
@@ -101,8 +101,15 @@ class PaymentMethodReader extends AbstractPremappingReader
             if (isset($this->connectionPremappingDictionary[$data['id']])) {
                 $uuid = $this->connectionPremappingDictionary[$data['id']]['destinationUuid'];
             }
+            if (!empty($data['description'])) {
+                $description = $data['description'];
+            } elseif (!empty($data['name'])) {
+                $description = $data['name'];
+            } else {
+                $description = $data['id'];
+            }
 
-            $entityData[] = new PremappingEntityStruct($data['id'], $data['description'], $uuid);
+            $entityData[] = new PremappingEntityStruct($data['id'], $description, $uuid);
         }
 
         $uuid = '';