+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - MIG-85 - Migrate customer comments in orders
 - MIG-90 - Fix variant migration
 - MIG-92 - Fix broken history download
+- MIG-98 - Fix premappings without description
 
 # 2.1.1
 - MIG-39 - Optimized basic converter